S_News/requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dataclasses==0.7
Django==2.0.5
django-multiselectfield==0.1.12
django-webpush==0.3.2
djangorestframework==3.8.0
fcm-django==0.3.4
http-ece==1.1.0
idna==2.9
mysqlclient==1.4.6
ndg-httpsclient==0.5.1
Pillow==7.0.0
pusher==3.0.0
pusher-push-notifications==2.0.0
py-vapid==1.7.0
pyasn1==0.4.8
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pytz==2019.3
pywebpush==1.9.4
requests==2.23.0
six==1.14.0
urllib3==1.25.9
